Add schema, testdata and boilerplate for build
Add the schema and data to drive the tsbench homework assignment. These
files have been provided as the starting point.

Add a Makefile and script to start Timescale DB in a docker container.

Also in the Makefile is the boilerplate to build, test, lint and check
test coverage of the Go program, soon to be added.