Add initial version of tsbench
The initial version of tsbench reads and validates an input CSV file,
counting the number of queries in the nascent data pipeline for
processing the input. Counting the queries is a temporary endpoint for
the pipeline, giving the program a running form and allowing it to
evolve.

Add tests for the code present, dropping the coverage threshold from
100% (for no code) to 80%, as `main()` is not covered.

Subsequent versions of tsbench will extend the pipeline until it
meets the requirements of the homework assignment.