fixup! Extend pipeline to query database and summarise results
Add functions for sending and receiving on the two channel types we use
that take a context to allow the send/receive to be interrupted. This
was open-coded before and made for a bit of a mess. This seems more
readable, although it may become a bit unwieldy if more channel types
get used. In that case, a generic version could be written with the
reflect package, or maybe just wait for generics to be added to the
language (any day now ;-) ).