Add concurrent workers
Make the query execution run in a worker pool, the size based on a
command line flag. Work is distributed to workers based on the FNV-1a
hash of the hostname.

Also print out the wall clock run time of the program so we can see the
comparison between the CPU time of all the workers vs how long it
actually takes to run.